Add display comment counter functionality
