+++ src/modules/setComment.js
@@ -34,16 +34,19 @@ const getComments = async (id) => {
     options,
   );
 
-  const comments = response.json();
-  return comments;
+  const comments = await response.json();
+  if (comments.length > 0) {
+    return comments;
+  } else {return []}
+  // return (comments.length > 0) ? comments : [];
 };
 
-const displayComment = (commentsArray, param, param1) => {
-  param.innerHTML = '';
-  commentsArray.forEach((element) => {
-    param.innerHTML += `<div>${element.creation_date} ${element.username}: ${element.comment}</div>`;
-    param1.innerHTML = `Total comment (${commentsArray.length})`;
-  });
-};
+// const displayComment = (commentsArray, param, param1) => {
+//   param.innerHTML = '';
+//   commentsArray.forEach((element) => {
+//     param.innerHTML += `<div>${element.creation_date} ${element.username}: ${element.comment}</div>`;
+//     param1.innerHTML = `Total comment (${commentsArray.length})`;
+//   });
+// };
 
-export { setComment, getComments, displayComment };
+export { setComment, getComments };
